Use output channel names referrring to destination
Channels derived from the output of processes were
renamed so that they correspond to their destination,
instead of the content that they are holding. This makes
it less confusing when a process can be skipped. E.g.,
previously, the output channel for trimmed fastp reads was
ch_trimmed_reads, and this was the input value expected by
the hostremoval process. But if trimming was skipped, the
channel would have the same name, but it would be holding
the raw reads.