Allow multi-lane samples
The collection of samples from the samplesheet now handles multi-lane samples the way sarek does. A unique id is constructed based on the sample name and lane id. Simple checks are run to check if samples are unique. Previous test for consistent paired/single-end status is retained.
Several todos added.