Fix missing fastp output reads channel
The `discard_trimmed_pass` parameter was erroneously set to
false, which prevented the .reads process output from containing
any reads.
This was caused by setting the parameter equal to
`!params.fastp_save_trimmed`, but that save_trimmed parameter
was intended to be used in the modules.config publishDir
declaration to enable/disable storing the output.